Warning fixes
The original code must have been compiled with a low warning
level. These fixes are to bring it into warning-free state
vis-a-vis VC141 /W4 /WX (max warnings on Visual Studio 2017).

+++ cat-file.c
@@ -17,7 +17,7 @@ int main(int argc, char **argv)
 	fd = mkstemp(template);
 	if (fd < 0)
 		usage("unable to create tempfile");
-	if (write(fd, buf, size) != size)
+	if (write(fd, buf, size) != (ssize_t) size)
 		strcpy(type, "bad");
 	printf("%s: %s\n", template, type);
 }
